BlockMacros: runtime helpers moved to BlockMacrosRuntime
